✅ 유저 생성 픽스쳐 분리
